+++ lib/hexapdf/type/resources.rb
@@ -222,6 +222,10 @@ module HexaPDF
           yield("No procedure set specified", true)
           self[:ProcSet] = [:PDF, :Text, :ImageB, :ImageC, :ImageI]
         else
+          if val.kind_of?(Symbol)
+            yield("Procedure set is a single value instead of an Array", true)
+            val = value[:ProcSet] = [val]
+          end
           val.reject! do |name|
             case name
             when :PDF, :Text, :ImageB, :ImageC, :ImageI